# cocktail-network-ui/src/app/services/models/chat-message.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ChatMessage {
  chatId?: string;
  content?: string;
  id?: string;
  recipientId?: string;
  senderId?: string;
  timestamp?: string;
}