# cocktail-network-ui/src/app/services/models/cocktail-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CocktailRequest {
  authorName: string;
  id?: number;
  recipe: string;
  title: string;
}