# cocktail-network-ui/src/app/services/models/cocktail-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CocktailResponse {
  authorName?: string;
  cover?: Array<string>;
  id?: number;
  owner?: string;
  rate?: number;
  recipe?: string;
  title?: string;
}