# cocktail-network-ui/src/app/services/models/cocktail.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Feedback } from '../models/feedback';
import { User } from '../models/user';
export interface Cocktail {
  authorName?: string;
  cover?: string;
  createdBy?: number;
  createdDate?: string;
  feedbacks?: Array<Feedback>;
  id?: number;
  lastModifiedBy?: number;
  lastModifiedDate?: string;
  owner?: User;
  rate?: number;
  recipe?: string;
  title?: string;
}