# cocktail-network/src/main/java/com/andrei/cocktail/feedback/FeedbackController.java
package com.andrei.cocktail.feedback;

import com.andrei.cocktail.common.PageResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("feedbacks")
@RequiredArgsConstructor
@Tag(name = "Feedback")
public class FeedbackController {

    private final FeedbackService service;

    @PostMapping
    public ResponseEntity<Integer>  saveFeedback(
            @Valid @RequestBody FeedbackRequest request,
            Authentication connectedUser
    ) {
        return ResponseEntity.ok(service.save(request, connectedUser));
    }

    @GetMapping("/cocktail/{cocktail-id}")
    public ResponseEntity<PageResponse<FeedbackResponse>> findAllFeedbacksByCocktail(
            @PathVariable("cocktail-id") Integer cocktailId,
            @RequestParam(name = "page", defaultValue = "0", required = false) int page,
            @RequestParam(name = "size", defaultValue = "10", required = false) int size,
            Authentication connectedUser
    ) {
        return ResponseEntity.ok(service.findAllFeedbacksByCocktail(cocktailId, page, size, connectedUser));
    }
}