# cocktail-network/src/main/java/com/andrei/cocktail/feedback/FeedbackResponse.java
package com.andrei.cocktail.feedback;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class FeedbackResponse {

    private Double note;
    private String comment;
    private boolean ownFeedback;
}